[cli] release script WIP
