+++ cli/CHANGELOG.md
@@ -1,0 +1,5 @@
+# Mops CLI Changelog
+
+## 0.39.2
+- Test
+- Test2
